refactor(my-events): extract utils to respect 150-line limit
Refatora useMyEvents.ts para respeitar regra de máximo 150 linhas por arquivo.

**Problema Identificado:**
- useMyEvents.ts tinha 210 linhas, ultrapassando limite de 150

**Solução:**
Extraídas funções para novos arquivos utils:

1. **eventShare.ts** (56 linhas)
   - Função `shareEvent()` com formatação de mensagem
   - Lógica de compartilhamento de eventos

2. **buildMyEventsFilters.ts** (25 linhas)
   - Função `buildMyEventsFilters()`
   - Mapeamento de EventFilterType para UserEventStatus

**Resultado:**
- useMyEvents.ts: 210 → 137 linhas ✅
- Todos arquivos < 150 linhas
- Melhor separação de responsabilidades
- Código mais testável e reutilizável

**Validação Completa:**
✅ 0 erros TypeScript (strict mode)
✅ 0 warnings ESLint
✅ Estrutura de arquivos: kebab-case (pastas), PascalCase (componentes)
✅ Nomenclatura: prefixos corretos (styles*, types*, use*)
✅ Todos arquivos < 150 linhas
✅ Tokens Arena (sem hardcoded values)
✅ Text components com variant obrigatória
✅ Separação de concerns (styles, types, hooks)
