+++ src/screens/myEventsScreen/hooks/useMyEvents.ts
@@ -1,8 +1,9 @@
 import { useState, useCallback, useRef } from 'react';
-import { Share } from 'react-native';
-import { Event, EventsFilter } from '@/services/events/typesEvents';
+import { Event } from '@/services/events/typesEvents';
 import { eventsService } from '@/services/events/eventsService';
 import { EventFilterType } from '../typesMyEventsScreen';
+import { buildMyEventsFilters } from '../utils/buildMyEventsFilters';
+import { shareEvent } from '../utils/eventShare';
 
 interface UseMyEventsParams {
   eventFilter: EventFilterType;
@@ -38,30 +39,6 @@ export const useMyEvents = ({
 
   eventFilterRef.current = eventFilter;
 
-  const buildApiFilters = useCallback(
-    (filter: EventFilterType): EventsFilter => {
-      const baseFilters: EventsFilter = {};
-
-      if (filter === 'participating') {
-        baseFilters.userEventStatus = ['PARTICIPANT'];
-      } else if (filter === 'invited') {
-        baseFilters.userEventStatus = ['INVITED'];
-      } else if (filter === 'organizing') {
-        baseFilters.userEventStatus = ['ORGANIZER', 'ADMIN'];
-      } else {
-        baseFilters.userEventStatus = [
-          'ORGANIZER',
-          'ADMIN',
-          'PARTICIPANT',
-          'INVITED',
-        ];
-      }
-
-      return baseFilters;
-    },
-    []
-  );
-
   const loadEvents = useCallback(async () => {
     if (isLoadingRef.current) {
       return;
@@ -72,7 +49,7 @@ export const useMyEvents = ({
       setIsLoading(true);
       setError(null);
 
-      const filters = buildApiFilters(eventFilterRef.current);
+      const filters = buildMyEventsFilters(eventFilterRef.current);
       const response = await eventsService.getFeedEvents(1, filters);
 
       setEvents(response.data);
@@ -86,7 +63,7 @@ export const useMyEvents = ({
       setIsLoading(false);
       isLoadingRef.current = false;
     }
-  }, [buildApiFilters]);
+  }, []);
 
   const loadMoreEvents = useCallback(async () => {
     if (isLoadingMore || !hasMore || isLoadingRef.current) return;
@@ -95,7 +72,7 @@ export const useMyEvents = ({
       setIsLoadingMore(true);
       const nextPage = currentPage + 1;
 
-      const filters = buildApiFilters(eventFilterRef.current);
+      const filters = buildMyEventsFilters(eventFilterRef.current);
       const response = await eventsService.getFeedEvents(nextPage, filters);
 
       setEvents(prev => [...prev, ...response.data]);
@@ -108,7 +85,7 @@ export const useMyEvents = ({
     } finally {
       setIsLoadingMore(false);
     }
-  }, [currentPage, hasMore, isLoadingMore, buildApiFilters]);
+  }, [currentPage, hasMore, isLoadingMore]);
 
   const refreshEvents = useCallback(async () => {
     if (isLoadingRef.current) return;
@@ -118,7 +95,7 @@ export const useMyEvents = ({
       setIsRefreshing(true);
       setError(null);
 
-      const filters = buildApiFilters(eventFilterRef.current);
+      const filters = buildMyEventsFilters(eventFilterRef.current);
       const response = await eventsService.getFeedEvents(1, filters);
 
       setEvents(response.data);
@@ -132,64 +109,14 @@ export const useMyEvents = ({
       setIsRefreshing(false);
       isLoadingRef.current = false;
     }
-  }, [buildApiFilters]);
+  }, []);
 
   const handleShare = useCallback(
     async (eventId: string) => {
       const event = events.find(e => e.id === eventId);
       if (!event) return;
 
-      try {
-        const eventDate = new Date(event.startDate);
-        const day = eventDate.getDate().toString().padStart(2, '0');
-        const month = eventDate.getMonth() + 1;
-        const monthNames = [
-          'janeiro',
-          'fevereiro',
-          'março',
-          'abril',
-          'maio',
-          'junho',
-          'julho',
-          'agosto',
-          'setembro',
-          'outubro',
-          'novembro',
-          'dezembro',
-        ];
-        const hours = eventDate.getHours().toString().padStart(2, '0');
-        const minutes = eventDate.getMinutes().toString().padStart(2, '0');
-        const formattedDate = `${day} de ${monthNames[month - 1]} às ${hours}:${minutes}`;
-
-        const price = event.isFree
-          ? 'Gratuito'
-          : `R$ ${
-              typeof event.price === 'number'
-                ? event.price.toFixed(2)
-                : event.price
-            }`;
-
-        const location = `${event.location.city}, ${event.location.state}`;
-
-        const message = `🏃 ${event.sport.name}: ${event.title}
-
-📅 ${formattedDate}
-📍 ${location}
-💰 ${price}
-👥 ${event.currentParticipants}/${event.maxParticipants} participantes
-
-${event.description ? `\n${event.description}\n` : ''}
-Participe pelo app Arena! 🔥`;
-
-        await Share.share({
-          message,
-          title: `Arena - ${event.title}`,
-        });
-      } catch (error) {
-        if (__DEV__) {
-          throw error;
-        }
-      }
+      await shareEvent(event);
     },
     [events]
   );
